Add files via upload
The connector.py file is the workhorse of this connector for Fivetran, and the configuration.json will need to be updated with you Bolt API key which you will find in the "Open API" section of the settings menu. Also worth noting that you will need to alter the Token name in the connector.py file to match yours, as well. Just look for "TOKEN_NAME" and replace it to match your companies ID.